Enemy "stuck" to gate corners fixed & update timer logic
Fixed the bug when player tries to camp at one of four corners of the map, enemies spawning tend to be stuck at the gate. (Added a frictionless materia;)
Move spawnpoint closer to the center point a litle bit for safety
Change the timer logic: When the time is up, spawning mechanic will disable, but player have to defeat all the enemy before proceeding to next area.
TimerFill now will change to red when it is at 10%